feat: add author attribution to social cards and migrate share page routing
